Add app delegate command (#606)
* Add app delegate command stub

* Add app registration check

* Add lifetime in seconds option

* Read key and UCAN from environment

* Add DID ParseError

* Add UCAN ParseError

* Add proof token check

* Add an error when delegating without UCAN env var

* Add an error when delegating without machine key

Both machine key and UCAN must be set to delegate using
environment variables.

* Print delegated UCAN on success

* Separate user config and env var proof checks

The check from user config might be a root credential, but the
environment variable will not. The user config check may get an index of
apps from the server.

* Check user config UCAN for capability

* Add check for remote option

Check for the remote to support staging, localhost, and other
remotes.

* Add quiet flag

Command only outputs the UCAN with this flag. Quiet mode makes it easier
to use UCANs in a non-interactive use cases.

* Add potency option

Potency can be AppendOnly, Destroy, or Super_User. Defaults to
AppendOnly.

* Change AppendOnly to Append in help and errors